requirements.txt
pytest~=8.3.3
selene==2.0.0rc7
selenium~=4.26.1
jsonschema
requests
allure-pytest
allure-python-commons~=2.13.5